refac: create classes for the old request path and the new request path
