--- Assets/Scripts/NavGridSystem/IAgent.cs
@@ -1,7 +1,0 @@
-namespace NavGridSystem
-{
-    public interface IAgent
-    {
-        void SetPath(Cell[] path);
-    }
-}
