--- Assets/Scripts/NavGridSystem/INavigation.cs
@@ -1,12 +1,0 @@
-using Unity.Collections;
-using UnityEngine;
-
-namespace NavGridSystem
-{
-    public interface INavigation
-    {
-        void RequestPath(ref NativeList<Cell> path, Vector3 start, Vector3 end);
-        void RequestPath(ref NativeList<Cell> path, Cell start, Cell end);
-        void RequestPath(IAgent agent, Cell start, Cell end);
-    }
-}
