+++ Assets/Scripts/Pathfinding/AStarJob.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: b7f398680cf841478d1db1384cf2a35c
+timeCreated: 1752175343
