+++ Assets/Scripts/Pathfinding/IPathRequest.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9b55d7e21cbd4be690962aae1e7df7dd
+timeCreated: 1752174091
