+++ Assets/Scripts/Pathfinding/IPathfinding.cs
@@ -1,0 +1,9 @@
+using NavigationGraph;
+
+namespace Pathfinding
+{
+    public interface IPathfinding
+    {
+        void RequestPath(IAgent agent, Cell start, Cell end);
+    }
+}
